+++ Task-2-2-1/src/main/java/ru/nsu/krasnyanskii/pizzeria/BlockingOrderQueue.java
@@ -10,13 +10,6 @@ import java.util.List;
  * <p>Uses manual synchronization because {@code java.util.concurrent.BlockingQueue}
  * is prohibited by the assignment.</p>
  *
- * <p><b>synchronized variants:</b>
- * (1) {@code synchronized} method — lock on the instance ({@code this});
- * (2) {@code synchronized} block — lock on an explicit object, narrows the critical
- * section or allows separate locks for independent data;
- * (3) {@code static synchronized} method — lock on the {@code Class} object, one lock
- * per class across the entire JVM regardless of instance count.</p>
- *
  * @param <T> element type
  */
 public class BlockingOrderQueue<T> {
@@ -41,12 +34,6 @@ public class BlockingOrderQueue<T> {
     /**
      * Inserts an element, blocking until space is available.
      *
-     * <p>{@link InterruptedException} is thrown when a thread blocked in
-     * {@code wait}/{@code sleep}/{@code join} receives {@code thread.interrupt()}.
-     * This is Java's cooperative cancellation mechanism — the thread must either
-     * re-throw the exception or restore the interrupted flag via
-     * {@code Thread.currentThread().interrupt()}.</p>
-     *
      * @param item element to insert
      * @throws InterruptedException if the thread is interrupted while waiting
      */
@@ -64,8 +51,6 @@ public class BlockingOrderQueue<T> {
     /**
      * Retrieves and removes the head element, blocking until one is available.
      *
-     * <p>For {@link InterruptedException} semantics see {@link #put(Object)}.</p>
-     *
      * @return the head element, or {@code null} if the queue is closed and empty
      * @throws InterruptedException if the thread is interrupted while waiting
      */
@@ -74,7 +59,7 @@ public class BlockingOrderQueue<T> {
             wait();
         }
         if (queue.isEmpty()) {
-            return null; // closed and empty — signal for workers to stop
+            return null;
         }
         T item = queue.removeFirst();
         notifyAll();
