+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/OrderSerializerTest.java
@@ -9,6 +9,9 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
+import ru.nsu.krasnyanskii.pizzeria.config.OrderSerializer;
+import ru.nsu.krasnyanskii.pizzeria.model.Order;
+import ru.nsu.krasnyanskii.pizzeria.view.PizzeriaView;
 
 class OrderSerializerTest {
