+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/PizzaStorageTest.java
@@ -8,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import ru.nsu.krasnyanskii.pizzeria.model.Order;
 
 class PizzaStorageTest {
