+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/PizzeriaConfigTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.ArrayList;
 import org.junit.jupiter.api.Test;
+import ru.nsu.krasnyanskii.pizzeria.model.PizzeriaConfig;
 
 class PizzeriaConfigTest {
