+++ Task-2-2-1/src/test/java/ru/nsu/krasnyanskii/pizzeria/PizzeriaMainTest.java
@@ -7,6 +7,7 @@ import java.nio.file.Path;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
+import ru.nsu.krasnyanskii.pizzeria.model.Order;
 
 import static org.junit.jupiter.api.Assertions.*;
